Move bitPositionsToArray/bitmaskToArray to separate file
